Added explore command
Changed callback function signature to take optional additional commands via the ...any parameter. Added logic in main() loop to check if additional parameters are given or not. Per command the additional parameter can be evaluated. For now, the explore command just calls the API and returns a 404 if the user entered a wrong location. If the explore command is missing an additional argument, no call is being made to API to indicate to user they have to specify a command. Changed the formatting of some of the returns on fmt.Printf for more legibility.